Replace XOR firmware encryption with Speck 32/64-CTR (within the 4 KB boot section)

+++ bootloader/src/bootloader_timer.h
@@ -4,7 +4,11 @@
 #include <stdint.h>
 #include <stdbool.h>
 
-typedef int32_t BT_time_ms_t;
+// 16 bit bastano: il timeout massimo (60000 ms) sta in uint16 e su AVR 'int' e'
+// a 16 bit, quindi l'aritmetica resta 16-bit unsigned modulare (meta' codice
+// rispetto a 32 bit). Il confronto in BT_timer_expired usa un cast esplicito
+// per garantire il wraparound corretto a 65536 ms.
+typedef uint16_t BT_time_ms_t;
 
 extern volatile BT_time_ms_t bootloader_timer;
